+++ src/tarmac/main.py
@@ -10,7 +10,7 @@ def main(args=None):
     logging.basicConfig(level=logging.INFO)
 
     parser = argparse.ArgumentParser(description="Execute a script with inputs")
-    parser.add_argument("job", type=str, help="The job to execute")
+    parser.add_argument("workflow", type=str, help="The workflow to execute")
     parser.add_argument(
         "-b",
         "--base-path",
@@ -38,6 +38,6 @@ def main(args=None):
         or os.environ.get("TARMAC_BASE_PATH", "")
         or os.getcwd()
     )
-    result = runner.execute_job(args.job, inputs)
+    result = runner.execute_workflow(args.workflow, inputs)
     print()
     print(yaml.dump(result, indent=2))