+++ test/test_main.py
@@ -2,7 +2,7 @@ import pytest
 from pathlib import Path
 
 
-def test_help(config_dir: Path) -> None:
+def test_help(config_dir: Path):
     from tarmac.main import main
 
     with pytest.raises(SystemExit):